[paywall] Move non-pages out of pages dir
